perf(DisplayImageUpload): code-split Uppy away
This makes Uppy load on demand (and shaves ~200 minified kBs from the bundle)
